Partial merge of ft3

+++ collect_app/src/main/java/org/odk/collect/android/provider/FormsProvider.java
@@ -48,7 +48,7 @@ public class FormsProvider extends ContentProvider {
 	private static final String t = "FormsProvider";
 
 	private static final String DATABASE_NAME = "forms.db";
-	private static final int DATABASE_VERSION = 4;
+	private static final int DATABASE_VERSION = 7;	// smap must be greater than 4
 	private static final String FORMS_TABLE_NAME = "forms";
 
 	private static HashMap<String, String> sFormsProjectionMap;
@@ -85,6 +85,12 @@ public class FormsProvider extends ContentProvider {
 					+ " text not null, "
 					+ FormsColumns.JR_VERSION
 					+ " text, "
+					+ FormsColumns.PROJECT		// smap
+					+ " text, "					// smap
+                    + FormsColumns.TASKS_ONLY	// smap
+                    + " text, "					// smap
+					+ FormsColumns.SOURCE		// smap
+					+ " text, "					// smap
 					+ FormsColumns.MD5_HASH
 					+ " text not null, "
 					+ FormsColumns.DATE
@@ -107,7 +113,9 @@ public class FormsProvider extends ContentProvider {
 				db.execSQL("DROP TABLE IF EXISTS " + FORMS_TABLE_NAME);
 				onCreate(db);
 				return;
-			} else {
+			} else if(oldVersion < 6) {
+				Log.w(t, "Upgrading database from version " + oldVersion		// smap
+						+ " to " + newVersion);
 				// adding BASE64_RSA_PUBLIC_KEY and changing type and name of
 				// integer MODEL_VERSION to text VERSION
 				db.execSQL("DROP TABLE IF EXISTS " + TEMP_FORMS_TABLE_NAME);
@@ -137,9 +145,13 @@ public class FormsProvider extends ContentProvider {
 						+ ", "
 						+ FormsColumns.SUBMISSION_URI
 						+ ", "
-						+ FormsColumns.JR_VERSION
-						+ ", "
-						+ ((oldVersion != 3) ? ""
+						+ ((oldVersion > 3) ? ""					// smap
+								: (FormsColumns.JR_VERSION + ", "))	// smap
+						+ ((oldVersion < 5) ? ""					// smap
+								: (FormsColumns.PROJECT + ", "))	// smap
+						+ ((oldVersion < 6) ? ""					// smap
+								: (FormsColumns.SOURCE + ", "))	// smap
+						+ ((oldVersion < 4) ? ""					// smap
 								: (FormsColumns.BASE64_RSA_PUBLIC_KEY + ", "))
 						+ FormsColumns.JRCACHE_FILE_PATH
 						+ ") SELECT "
@@ -165,13 +177,20 @@ public class FormsProvider extends ContentProvider {
 						+ ", "
 						+ FormsColumns.SUBMISSION_URI
 						+ ", "
-						+ "CASE WHEN "
-						+ MODEL_VERSION
-						+ " IS NOT NULL THEN "
-						+ "CAST("
-						+ MODEL_VERSION
-						+ " AS TEXT) ELSE NULL END, "
-						+ ((oldVersion != 3) ? ""
+						+ ((oldVersion > 3) ? ""							// smap
+								: (
+										"CASE WHEN "		// smap
+										+ MODEL_VERSION
+										+ " IS NOT NULL THEN "
+										+ "CAST("
+										+ MODEL_VERSION
+										+ " AS TEXT) ELSE NULL END, "
+								))
+						+ ((oldVersion < 5) ? ""						// smap
+								: (FormsColumns.PROJECT + ", "))		// smap
+						+ ((oldVersion < 6) ? ""						// smap
+								: (FormsColumns.SOURCE + ", "))			// smap
+						+ ((oldVersion < 4) ? ""						// smap
 								: (FormsColumns.BASE64_RSA_PUBLIC_KEY + ", "))
 						+ FormsColumns.JRCACHE_FILE_PATH + " FROM "
 						+ FORMS_TABLE_NAME);
@@ -201,6 +220,8 @@ public class FormsProvider extends ContentProvider {
 						+ FormsColumns.LANGUAGE + ", "
 						+ FormsColumns.SUBMISSION_URI + ", "
 						+ FormsColumns.JR_VERSION + ", "
+						+ FormsColumns.PROJECT + ", "				// smap
+						+ FormsColumns.SOURCE + ", "				// smap
 						+ FormsColumns.BASE64_RSA_PUBLIC_KEY + ", "
 						+ FormsColumns.JRCACHE_FILE_PATH + ") SELECT "
 						+ FormsColumns._ID + ", "
@@ -221,6 +242,8 @@ public class FormsProvider extends ContentProvider {
 						+ FormsColumns.LANGUAGE + ", "
 						+ FormsColumns.SUBMISSION_URI + ", "
 						+ FormsColumns.JR_VERSION + ", "
+						+ FormsColumns.PROJECT + ", "				// smap
+						+ FormsColumns.SOURCE + ", "				// smap
 						+ FormsColumns.BASE64_RSA_PUBLIC_KEY + ", "
 						+ FormsColumns.JRCACHE_FILE_PATH + " FROM "
 						+ TEMP_FORMS_TABLE_NAME);
@@ -230,6 +253,19 @@ public class FormsProvider extends ContentProvider {
 						+ initialVersion + " to " + newVersion
 						+ ", without destroying all the old data");
 			}
+
+            if ( oldVersion < 7 ) {
+                try {
+                    db.execSQL("ALTER TABLE " + FORMS_TABLE_NAME + " ADD COLUMN " +
+                            FormsColumns.TASKS_ONLY + " text;");
+                    db.execSQL("update " + FORMS_TABLE_NAME + " set " +
+                            FormsColumns.TASKS_ONLY + " = 'no';" );
+                }catch(Exception e) {
+                    // Catch errors, its possible the user upgraded then downgraded
+                    Log.w(t, "Error in upgrading to forms database version 7");
+                    e.printStackTrace();
+                }
+            }
 		}
 	}
 
@@ -712,6 +748,12 @@ public class FormsProvider extends ContentProvider {
 				FormsColumns.JR_FORM_ID);
 		sFormsProjectionMap.put(FormsColumns.JR_VERSION,
 				FormsColumns.JR_VERSION);
+		sFormsProjectionMap.put(FormsColumns.PROJECT,		// smap
+				FormsColumns.PROJECT);						// smap
+        sFormsProjectionMap.put(FormsColumns.TASKS_ONLY,	// smap
+                FormsColumns.TASKS_ONLY);                   // smap
+        sFormsProjectionMap.put(FormsColumns.SOURCE,		// smap
+				FormsColumns.SOURCE);						// smap
 		sFormsProjectionMap.put(FormsColumns.SUBMISSION_URI,
 				FormsColumns.SUBMISSION_URI);
 		sFormsProjectionMap.put(FormsColumns.BASE64_RSA_PUBLIC_KEY,
